pyproject.toml
[tool.poetry]
name = "app-albion-accounting"
version = "0.1.0"
description = ""
authors = ["Eric Batista"]
readme = "README.md"
packages = [{include = "app_albion_accounting"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.16.0"
black = "^22.12.0"
isort = "^5.11.4"
autoflake = "^2.0.0"
uvicorn = "^0.20.0"
orjson = "^3.8.6"
fastapi = "^0.92.0"
eric-devtools = "^0.2.1"
aiohttp = "^3.8.4"
alembic = "^1.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
